refactor: move reactions api to app dir
